569: Moved SwapInputs to components
