+++ src/components/SwapInputs/helpers/getSwapInputIcon.tsx
