+++ src/components/SwapInputs/helpers/getTokenMaxInfoText.ts
